--- src/Components/PinnedNote/PinnedNote.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const PinnedNote = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export { PinnedNote}
