--- src/Components/RightAside/RightAside.css
@@ -1,99 +1,0 @@
-
-.layout__right-sidebar {
-  position: fixed;
-  width: 350px;
-  margin-left: 2rem;
-  z-index: 2;
-}
-
-.trends-for-you {
-  border-radius: 15px;
-  background: #f5f8fa;
-  margin-bottom: 15px;
-}
-
-.trends-for-you__block {
-  border-bottom: 1px solid #e6ecf0;
-  padding: 10px 15px;
-}
-
-.trends-for-you__block:last-child {
-  border-bottom: none;
-}
-
-.trends-for-you__heading {
-  font-size: 19px;
-  font-weight: bold;
-  color: #14171a;
-}
-
-.trends-for-you__meta-information {
-  font-size: 13px;
-  font-weight: 300;
-  color: #657786;
-}
-
-.trends-for-you__trend-name {
-  font-size: 15px;
-  color: #14171a;
-  font-weight: bold;
-  margin-bottom: 5px;
-  margin-top: 2px;
-}
-
-/* Who to follow css */
-
-
-.who-to-follow {
-  background: #f5f8fa;
-  border-radius: 15px;
-}
-
-
- .who-to-follow__block {
-  padding: 10px 15px;
-  display: flex;
-  align-items: center;
-  border-bottom: 1px solid #e6ecf0;
-}
-.who-to-follow__heading {
-  font-size: 19px;
-  font-weight: bold;
-  color: #14171a;
-}
-.who-to-follow__block:last-child {
-  border-bottom: none;
-}
-
-/* searchbox */
-.layout__right-sidebar-container-head{
-    display: flex;
-    justify-content: flex-start;
-    padding: 1rem 0 1rem 2rem;
-    width: 350px;
-}
-.nav-bottom-search-box {
-    align-self: flex-end;
-    font-size: medium;
-    border: var(--GREY) solid;
-    border-radius: 1.7rem;
-    padding: 0 1rem;
-    box-shadow: inset 0 1px 1px rgb(0 0 0 / 8%);
-    margin-right: 1rem;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.nav-bottom-search {
-    width: 90%;
-    height: 100%;
-    font-size: medium;
-    border: none;
-    background: url(../../Assets/Icons/pngwing.com.png) center right no-repeat;
-}
-
-.nav-bottom-search:focus {
-    outline: none;
-}
